requirements.txt
asgiref==3.8.1
Django==5.1.2
django-axes==7.0.0
django-crispy-forms==2.3
django-filter==24.3
django-formtools==2.5.1
django-guardian==2.4.0
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.7
packaging==24.1
phonenumbers==8.13.47
pillow==10.4.0
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
